Move cty-run to `cty-run`.
